app/requirements.txt
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
Flask==2.2.2
future==0.18.3
gunicorn==20.1.0
idna==3.4
impacket==0.10.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==2.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pyOpenSSL==23.0.0
python-ldap==3.4.3
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==2.2.2
zipp==3.12.1